requirements.txt
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
factory-boy==3.3.0
